+++ appJs/appHtmlJs/src/commonMain/kotlin/com/gchristov/thecodinglove/Readme.kt
@@ -1,3 +1,5 @@
+package com.gchristov.thecodinglove
+
 /*
  * The website still needs a source file in order to compile correctly. We use a placeholder since
  * There isn't any functionality needed yet.
