+++ gradle-plugins/build.gradle.kts
@@ -8,15 +8,14 @@ gradlePlugin {
         implementationClass =
             "com.gchristov.thecodinglove.gradleplugins.JavascriptBrowserExecutablePlugin"
     }
-    plugins.register("javascript-node-executable-plugin") {
-        id = "javascript-node-executable-plugin"
+    plugins.register("javascript-node-library-plugin") {
+        id = "javascript-node-library-plugin"
         implementationClass =
-            "com.gchristov.thecodinglove.gradleplugins.JavascriptNodeExecutablePlugin"
+            "com.gchristov.thecodinglove.gradleplugins.JavascriptNodeLibraryPlugin"
     }
-    plugins.register("javascript-node-platform-plugin") {
-        id = "javascript-node-platform-plugin"
-        implementationClass =
-            "com.gchristov.thecodinglove.gradleplugins.JavascriptNodePlatformPlugin"
+    plugins.register("javascript-node-target-plugin") {
+        id = "javascript-node-target-plugin"
+        implementationClass = "com.gchristov.thecodinglove.gradleplugins.JavascriptNodeTargetPlugin"
     }
     plugins.register("javascript-node-module-plugin") {
         id = "javascript-node-module-plugin"
