+++ kmp/kmp-common-di/build.gradle.kts
@@ -1,14 +1,14 @@
 import com.gchristov.thecodinglove.gradleplugins.Deps
 
 plugins {
-    id("javascript-node-platform-plugin")
+    id("javascript-node-target-plugin")
 }
 
 kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                implementation(Deps.Kodein.di)
+                api(Deps.Kodein.di)
             }
         }
     }
