+++ kmp/kmp-module-b/build.gradle.kts
@@ -1,13 +1,3 @@
 plugins {
     id("javascript-node-module-plugin")
 }
-
-kotlin {
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                implementation("org.kodein.di:kodein-di:7.15.0")
-            }
-        }
-    }
-}
